Initial commit of everything working!
This isn't perfect, but everything works as expected. I did have to
hard code the version of @hapi/boom for now due to an issue with the
packages. @hapi/boom updated, but the types for @hapi/hapi are not up
to date to allow the project to compile. Forcing the project to use a
previous version of the package avoids the issue for now