Turn off coverage; add helper method for HTTP GET a route
The coverage was just annoying. I can add that back if I want it
locally later, or when I eventually start using a CI tool.

The helper method clears up some boilerplate with using the server. I
thought about wrapping `start` up in it, and keeping a reference to
the server somewhere, but it's also kind of nice to see the setup for
the server local in the file. We'll see if it is still nice or becomes
annoying as the tests become more complicated.